requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
cacheout==0.11.1
chardet==3.0.4
discord.py==0.16.10
diskcache==4.0.0
multidict==4.5.2
PyYAML==3.11
requests==2.10.0
schedule==0.3.2
websockets==3.4
